+++ tomcat-base/src/shared/java/eu/xenit/alfresco/tomcat/embedded/config/DefaultConfigurationProvider.java
@@ -27,24 +27,24 @@ public class DefaultConfigurationProvider implements ConfigurationProvider {
         baseConfiguration.setGlobalProperty("db.password", "admin");
         baseConfiguration.setGlobalProperty("db.url", "jdbc:postgresql://${db.host}:${db.port}/${db.name}");
 
-        baseConfiguration.setGlobalProperty("encryption.keystore.type", "JCEKS");
-        baseConfiguration.setGlobalProperty("encryption.cipherAlgorithm", "DESede/CBC/PKCS5Padding");
-        baseConfiguration.setGlobalProperty("encryption.keyAlgorithm", "DESede");
-        baseConfiguration.setGlobalProperty("encryption.keystore.location", "/keystore/keystore");
-        baseConfiguration.setGlobalProperty("metadata-keystore.password", "mp6yc0UD9e");
-        baseConfiguration.setGlobalProperty("metadata-keystore.aliases", "metadata");
-        baseConfiguration.setGlobalProperty("metadata-keystore.metadata.password", "oKIWzVdEdA");
-        baseConfiguration.setGlobalProperty("metadata-keystore.metadata.algorithm", "DESede");
+        baseConfiguration.setSystemProperty("encryption.keystore.type", "JCEKS");
+        baseConfiguration.setSystemProperty("encryption.cipherAlgorithm", "DESede/CBC/PKCS5Padding");
+        baseConfiguration.setSystemProperty("encryption.keyAlgorithm", "DESede");
+        baseConfiguration.setSystemProperty("encryption.keystore.location", "/keystore/keystore");
+        baseConfiguration.setSystemProperty("metadata-keystore.password", "mp6yc0UD9e");
+        baseConfiguration.setSystemProperty("metadata-keystore.aliases", "metadata");
+        baseConfiguration.setSystemProperty("metadata-keystore.metadata.password", "oKIWzVdEdA");
+        baseConfiguration.setSystemProperty("metadata-keystore.metadata.algorithm", "DESede");
 
-        baseConfiguration.setGlobalProperty("encryption.ssl.keystore.location", "/keystore/ssl.keystore");
-        baseConfiguration.setGlobalProperty("encryption.ssl.keystore.type", "JCEKS");
-        baseConfiguration.setGlobalProperty("encryption.ssl.keystore.keyMetaData.location", "");
-        baseConfiguration.setGlobalProperty("encryption.ssl.truststore.location", "/keystore/ssl.truststore");
-        baseConfiguration.setGlobalProperty("encryption.ssl.truststore.type", "JCEKS");
-        baseConfiguration.setGlobalProperty("encryption.ssl.truststore.keyMetaData.location", "");
+        baseConfiguration.setSystemProperty("encryption.ssl.keystore.location", "/keystore/ssl.keystore");
+        baseConfiguration.setSystemProperty("encryption.ssl.keystore.type", "JCEKS");
+        baseConfiguration.setSystemProperty("encryption.ssl.keystore.keyMetaData.location", "");
+        baseConfiguration.setSystemProperty("encryption.ssl.truststore.location", "/keystore/ssl.truststore");
+        baseConfiguration.setSystemProperty("encryption.ssl.truststore.type", "JCEKS");
+        baseConfiguration.setSystemProperty("encryption.ssl.truststore.keyMetaData.location", "");
 
-        baseConfiguration.setGlobalProperty("ssl-truststore.password", "kT9X6oe68t");
-        baseConfiguration.setGlobalProperty("ssl-keystore.password", "kT9X6oe68t");
+        baseConfiguration.setSystemProperty("ssl-truststore.password", "kT9X6oe68t");
+        baseConfiguration.setSystemProperty("ssl-keystore.password", "kT9X6oe68t");
 
         baseConfiguration.setGlobalProperty("messaging.subsystem.autoStart", "false");
         baseConfiguration.setGlobalProperty("events.subsystem.autoStart", "false");
@@ -54,6 +54,8 @@ public class DefaultConfigurationProvider implements ConfigurationProvider {
         baseConfiguration.setGlobalProperty("solr.host", "solr");
         baseConfiguration.setGlobalProperty("solr.port", "8080");
         baseConfiguration.setGlobalProperty("solr.port.ssl", "8443");
+
+        baseConfiguration.setExitOnFailure(true);
 
         return baseConfiguration;
     }